graph toolkit basic ui

+++ src/main/java/com/lowdragmc/lowdraglib2/gui/ui/ModularUI.java
@@ -47,7 +47,7 @@ import org.appliedenergistics.yoga.YogaConstants;
 import org.joml.Matrix4f;
 import org.lwjgl.glfw.GLFW;
 
-import javax.annotation.Nullable;
+import org.jetbrains.annotations.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
@@ -1126,7 +1126,7 @@ public class ModularUI {
 
             // Do not render tooltips if carried item is existing
             if (drawDrag && dragHandler.isDragging() && dragHandler.dragTexture != null) {
-                dragHandler.dragTexture.draw(guiGraphics, (int) lastMouseX, (int) lastMouseY, lastMouseX + dragHandler.offsetX, lastMouseY + dragHandler.offsetY, dragHandler.width, dragHandler.height, partialTick);
+                dragHandler.dragTexture.draw(guiGraphics, lastMouseX, lastMouseY, lastMouseX + dragHandler.offsetX, lastMouseY + dragHandler.offsetY, dragHandler.width, dragHandler.height, partialTick);
             }
 
             if (drawTooltips && !dragHandler.isDragging() && tooltipTexts != null && !tooltipTexts.isEmpty()) {
